remove and get_locations done
