+++ lmcache/protocol.py
@@ -9,6 +9,7 @@ class Constants:
     CLIENT_GET = 2
     CLIENT_EXIST = 3
     CLIENT_LIST = 4
+    CLIENT_REMOVE = 5
 
     SERVER_SUCCESS = 200
     SERVER_FAIL = 400
